refactoring api
