+++ src/main/java/ru/yandex/school/hlebushek/Server.java
@@ -7,7 +7,7 @@ import javax.ws.rs.container.DynamicFeature;
 import javax.ws.rs.container.ResourceInfo;
 import javax.ws.rs.core.FeatureContext;
 
-import ru.yandex.school.hlebushek.service.DatabaseProvider;
+import ru.yandex.school.hlebushek.db.DatabaseProvider;
 
 public class Server extends ResourceConfig {
